Avoid deepcopying loader when cloning entities
Originally, the loader was not an attribute of an entity. It was added later as a performance optimization, so that all entities could share the same code cache for compiled fragments. Ironically, it has had the opposite effect, resulting in exponential memory use. This is bebecause the existing code for deep cloning, was now inadvertently deep-cloning the loader too, resulting in exponential memory use. This PR makes sure the loader alone is shallow-copied when deep-copying entities.