fix resubmission tests
on the 2nd submission the env variables get lost. For some reason
bash evaluates this as successful even if the if statement crashes.
so actually there are only 2 submissions (1 resubmission), i.e.
multiple resubmissions still do not work.

Hence I suggest to to a "real OOM" resubmission and treat the case
of a lost variable as error.